+++ src/oci/data_science/models/notebook_session_shape_summary.py
@@ -28,6 +28,10 @@ class NotebookSessionShapeSummary(object):
     SHAPE_SERIES_NVIDIA_GPU = "NVIDIA_GPU"
 
     #: A constant which can be used with the shape_series property of a NotebookSessionShapeSummary.
+    #: This constant has a value of "GENERIC"
+    SHAPE_SERIES_GENERIC = "GENERIC"
+
+    #: A constant which can be used with the shape_series property of a NotebookSessionShapeSummary.
     #: This constant has a value of "LEGACY"
     SHAPE_SERIES_LEGACY = "LEGACY"
 
@@ -54,7 +58,7 @@ class NotebookSessionShapeSummary(object):
 
         :param shape_series:
             The value to assign to the shape_series property of this NotebookSessionShapeSummary.
-            Allowed values for this property are: "AMD_ROME", "INTEL_SKYLAKE", "NVIDIA_GPU", "LEGACY", "ARM", 'UNKNOWN_ENUM_VALUE'.
+            Allowed values for this property are: "AMD_ROME", "INTEL_SKYLAKE", "NVIDIA_GPU", "GENERIC", "LEGACY", "ARM", 'UNKNOWN_ENUM_VALUE'.
             Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
         :type shape_series: str
 
@@ -156,7 +160,7 @@ class NotebookSessionShapeSummary(object):
         **[Required]** Gets the shape_series of this NotebookSessionShapeSummary.
         The family that the compute shape belongs to.
 
-        Allowed values for this property are: "AMD_ROME", "INTEL_SKYLAKE", "NVIDIA_GPU", "LEGACY", "ARM", 'UNKNOWN_ENUM_VALUE'.
+        Allowed values for this property are: "AMD_ROME", "INTEL_SKYLAKE", "NVIDIA_GPU", "GENERIC", "LEGACY", "ARM", 'UNKNOWN_ENUM_VALUE'.
         Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
 
 
@@ -175,7 +179,7 @@ class NotebookSessionShapeSummary(object):
         :param shape_series: The shape_series of this NotebookSessionShapeSummary.
         :type: str
         """
-        allowed_values = ["AMD_ROME", "INTEL_SKYLAKE", "NVIDIA_GPU", "LEGACY", "ARM"]
+        allowed_values = ["AMD_ROME", "INTEL_SKYLAKE", "NVIDIA_GPU", "GENERIC", "LEGACY", "ARM"]
         if not value_allowed_none_or_none_sentinel(shape_series, allowed_values):
             shape_series = 'UNKNOWN_ENUM_VALUE'
         self._shape_series = shape_series