+++ src/oci/data_science/models/pipeline_ml_job_step_details.py
@@ -23,7 +23,7 @@ class PipelineMLJobStepDetails(PipelineStepDetails):
 
         :param step_type:
             The value to assign to the step_type property of this PipelineMLJobStepDetails.
-            Allowed values for this property are: "ML_JOB", "CUSTOM_SCRIPT"
+            Allowed values for this property are: "ML_JOB", "CUSTOM_SCRIPT", "CONTAINER"
         :type step_type: str
 
         :param step_name: