+++ src/oci/data_science/models/pipeline_step_update_details.py
@@ -23,6 +23,10 @@ class PipelineStepUpdateDetails(object):
     #: This constant has a value of "CUSTOM_SCRIPT"
     STEP_TYPE_CUSTOM_SCRIPT = "CUSTOM_SCRIPT"
 
+    #: A constant which can be used with the step_type property of a PipelineStepUpdateDetails.
+    #: This constant has a value of "CONTAINER"
+    STEP_TYPE_CONTAINER = "CONTAINER"
+
     def __init__(self, **kwargs):
         """
         Initializes a new PipelineStepUpdateDetails object with values from keyword arguments. This class has the following subclasses and if you are using this class as input
@@ -30,12 +34,13 @@ class PipelineStepUpdateDetails(object):
 
         * :class:`~oci.data_science.models.PipelineMLJobStepUpdateDetails`
         * :class:`~oci.data_science.models.PipelineCustomScriptStepUpdateDetails`
+        * :class:`~oci.data_science.models.PipelineContainerStepUpdateDetails`
 
         The following keyword arguments are supported (corresponding to the getters/setters of this class):
 
         :param step_type:
             The value to assign to the step_type property of this PipelineStepUpdateDetails.
-            Allowed values for this property are: "ML_JOB", "CUSTOM_SCRIPT"
+            Allowed values for this property are: "ML_JOB", "CUSTOM_SCRIPT", "CONTAINER"
         :type step_type: str
 
         :param step_name:
@@ -83,6 +88,9 @@ class PipelineStepUpdateDetails(object):
 
         if type == 'CUSTOM_SCRIPT':
             return 'PipelineCustomScriptStepUpdateDetails'
+
+        if type == 'CONTAINER':
+            return 'PipelineContainerStepUpdateDetails'
         else:
             return 'PipelineStepUpdateDetails'
 
@@ -92,7 +100,7 @@ class PipelineStepUpdateDetails(object):
         **[Required]** Gets the step_type of this PipelineStepUpdateDetails.
         The type of step.
 
-        Allowed values for this property are: "ML_JOB", "CUSTOM_SCRIPT"
+        Allowed values for this property are: "ML_JOB", "CUSTOM_SCRIPT", "CONTAINER"
 
 
         :return: The step_type of this PipelineStepUpdateDetails.
@@ -110,7 +118,7 @@ class PipelineStepUpdateDetails(object):
         :param step_type: The step_type of this PipelineStepUpdateDetails.
         :type: str
         """
-        allowed_values = ["ML_JOB", "CUSTOM_SCRIPT"]
+        allowed_values = ["ML_JOB", "CUSTOM_SCRIPT", "CONTAINER"]
         if not value_allowed_none_or_none_sentinel(step_type, allowed_values):
             raise ValueError(
                 f"Invalid value for `step_type`, must be None or one of {allowed_values}"